Move the hero directory into the home directory
