+++ src/components/home/hero/DemoDialog.tsx
@@ -7,7 +7,7 @@ import {
     Stack,
     Typography,
 } from '@mui/material';
-import AcceptDemoInvitation from 'components/hero/AcceptDemoInvitation';
+import AcceptDemoInvitation from 'components/home/hero/AcceptDemoInvitation';
 import Logo from 'components/navigation/Logo';
 import { Dispatch, SetStateAction, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
