+++ src/components/home/hero/Detail.tsx
