+++ src/components/home/hero/Step.tsx
