+++ src/components/home/hero/Steps/Detail.tsx
