+++ src/components/home/hero/hooks.ts
