+++ src/components/home/hero/index.tsx
@@ -1,5 +1,5 @@
 import { Box, Grid } from '@mui/material';
-import HeroTabs from 'components/hero/Tabs';
+import HeroTabs from 'components/home/hero/Tabs';
 import HeroDemo from './Demo';
 import HeroDetail from './Detail';
 import { useHeroTabs } from './hooks';
