+++ src/components/home/hero/types.ts
